refactor aws client boot code
